requirement.txt
beautifulsoup4==4.12.3
requests==2.32.3
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.1
redis==5.0.8